LT-14704: Do not show TE Styles Dialog in FLEx
This class is used specifically for TE styles; FLEx *should* use a
different class, but per LT-14704, that is not the case. So always
checking for current styles, but suppressing a potentially confusing
dialog when TE is not installed.  (I was unable to
reproduce this locally, but am confident in this solution).

Change-Id: I5af9fb430869603527c75eeea9c78f6cd79a5190