Fix weird problem with manifest file
This change fixes a problem that suddently appeared on my machine:
when starting FW in VS FieldWorks.exe gets recompiled which caused
the manifest file to rebuild. For whatever reason registering
Language.dll failed which caused all comClass entries to be
removed from the manifest file, preventing FW from starting.

This change modifies the behavior of the RegFree.targets file that
gets included in the FieldWorks.csproj file: the manifest file gets
only rebuild if any of the unmanaged dlls changed.

Change-Id: I75ff9250794691bc4155e24e3221afceec8fa71f