FWNX-1146 Sorting now properly follows drag-n-drop
FWNX-1146
The issue problem was that dragging a column in any
browse view changed which column is sorted. Now the correct
column is being sorted.

Column order in the list is not actually being changed so an
index of the current position of columns has been created now
to ensure the icon is put on the correct column

Change-Id: I295333e68fedb564fb0fe1f24dda742ffcbc0e61