Fix parts 2 (of 2) of LT-14809 and LT-14836
The first part of the fix for LT-14809 is in Flex Bridge.
This second part of the fix (in Flex) handles the deleted
lift folder better.

The first part of the fix for LT-14836 is in Flex bridge.
Both parts aim to allow the Db4o BEP to work for Lift S/R,
as was done in the original Lift Bridge system.

Change-Id: I5761587f12670a7a155074f2e051eb5dae4f064a