Find Test Manifest file more reliably
Location of the FieldWorks.Tests.manifest file, required to avoid
having to register our COM DLLs, previously depended on the test
runner setting the current directory to the one containing our DLLs.
The Resharper 8 test runner does not do this, so I have made it
explicitly look for this file in the directory containing the
executing assembly file.

Change-Id: I721304831831cff43c1455ceec153852d0f5b130