FWNX-1206: make FileStreamXmlResolver really work
This class replaces the DoNothingXmlResolver, which turned out
not to do enough.

It now works well enough to handle at least simple cases on Linux.
Windows continues to just pass null for the resolver.

In order to verify this I had to fix another bug where export was
not working for vernacular sort rules other than toolbox or ICU.

Change-Id: Id0e7d17ccc4013a51fbed6d6d9bbe180e3386185