LT-14853 FW LDML reader sets WindowsLcid
The main point of this change is that the FieldWorks code
that reads LDML files needs to set the (legacy) WindowsLcid
field that the Palaso code uses to figure out the active keyboard.

Turns out the PalasoWritingSystem.LCID property is made obsolete
by this, so I removed it. Unfortunately it was inherited from
a COM interface, so that generated a lot of changes.

Change-Id: I853c113469a31c2239c6d7e32851b43a819f5a4b