FWNX-1233 don't dispose thread helper before cache
When obtaining a project, we were disposing of the thread helper too
soon, since it is a member variable of the cache, and used in the
import process which follows the end of the previous using block.

The only change (except to indentation and comments) is extending the
scope of the using block for the thread helper.

Change-Id: If2f927274cf170b07d5326d8de7697824857d87a