LT-14891 Matching morpheme in Sandbox copies all WSs
When we change the morpheme breakdown in the Sandbox, we try
to match the new morphemes to lex entries. When successful, if
we are displaying multiple WSs of the morphemes line, we need
to fill them in with the corresponding forms of the matched
morphs. Otherwise, confirming the analysis will delete those
alternatives from the morphemes.

Change-Id: I5af7212f1c02e2cd05a5e24c551fc77f57a5ab8a