LT-14922 Make LabeledMultiStringView an xCore colleague
When we added the layer of wrapping around the labelledMultiStringView
which is now the m_innerView of this class, the outer layer, which
is the actual control of the slice, was no longer a view and thus
no longer an IXCoreColleague. Consequently, the wrapped view was no
longer one of the message targets for its slice, and never given a
chance to enable various menu items.

Change-Id: I536f78366e90e1564c47785d32504b8656bbf859