The analytics was often reporting first time launches twice
There is a bug in the .Net settings class (I've reported it to MS)
If you save settings with a new sub object like ReportingSettings,
and then change that sub object and save again the changes aren't
written out to disk. The first save in our code was unnecessary so
removing it will avoid the bug.

Change-Id: Ib891da54418c678ba9fe2776486e4ee1dfb4d42e