Reinstate shutting down xpcom (FWNX-1216)
Not shutting down Xpcom seems to leave FieldWorks hanging around after
all the windows close.  Having an open web browser when it shuts down
seems to quell the "double free or corruption" message.  (This sort of
makes sense since the normal use of gecko is for a web browser that
always has a web browser in its main window.)

Change-Id: Ief0a366e38f8e35f7e63ff31741437c31b52eb72