LT-14503 Fix nesting of paragraphs and elements
Old DisplayVec code might, based on certain configuration attrs,
wrap a paragraph around each item in a sequence. However, under
certain circumstances, that paragraph would get broken inside the
display of the item. This led to improper nesting, a particular
problem for generated XML. Properly, the paragraph and anything
inside it is part of the display of a particular sequence item,
and should be nested inside the item display.
This code finds a rather kludgy way to make this so.

Change-Id: I6f3862c1c36924c7ef964f786611f2117f57959f