LT-14273 Make Messages slice show Project conflicts
Two parts to this, both depending on Chorus changes which you can
check in the branch NotesBarExtensions. One part passes additional
target guids, so that we show notes not just about the current entry
but about all the things it owns as well (e.g., we want to see
a note about a conflict in the gloss of a sense). The second part
deals with the fact that FLEx conflicts occur in files
Linguistics/Lexicon/*.lexdb.ChorusNotes and use the 'guid'
parameter in their URLs to identify the target object. (Regular notes,
including ones copied from the LIFT repo, live in
Lexicon.fwstub.ChorusNotes and use the 'id' parameter.)

To implement passing the guids of all owned objects I had to
also implement new FDO methods OwnedObjects and AllOwnedObjects.

Change-Id: I891fd8669810d499ba6f00b8f22e742ce683ed38