FWNX-915 Improve OnSetInitialContentObject handling
The problem here is a strange default component that is made the
main window content when the AreaListener's
OnSetInitialContentObject method is not called and the xWindow
one is called instead. Could not confirm the reason this happens
(or used to happen) only sometimes, but it MIGHT be related to the
arbitrary ordering of xCore colleagues at the same priority.
So one change is to make this more predictable.

I've also made the default behavior in xWindow throw an exception
which if it ever happens should give us some useful information
about what colleagues are active and how they are sorted.

Change-Id: I2a9b9a50ddaae30b7dc1382fa3db91daf5b13a07