FWNX-1235: Make FLEx see Paratext texts in Linux
This is done through the ParatextShared.dll, borrowed from PT
itself. Unfortunately the PT 7.3 Windows version we borrowed
until now does not work on Linux, and the Linux 7.3 dll is half-way
to 7.4 and has a changed API. When I got it more-or-less working
using the 7.3 Linux DLLs, one PT 7.4 project did not import.
Seemed best to upgrage to use the 7.4 DLLs.

Some other DLLs that ParatextShared depends on are added or modified.
The versions checked in here were obtained from Tom Hindle. They are
basically the Linux 7.4 beta 10 DLLs, with one fix to work around
an incompatibility between Mono and Windows handling of initializing
an array of enums, and others to make the Paratext code not require
Mono to be set to ignore case in file names. The source for these
DLLs is in the temporary branch 7.4-Linux-FlexChanges of the Paratext
source. These changes are not in either branch of PT 7.4 but are
expected to be in 7.5.

Change-Id: I18ab6ddc1d58108d9a90a79959843299d59d1b67