Fix keyboarding problem
When the user selects a different writing system in a multi-text
field we need to switch the writing system and the keyboard.
With IBus keyboards this didn't work previously because of the
way we made the selection in IbusRootSiteEventHandler. This caused
the writing to be lost so that we ended up with the surrounding
writing system instead of the WS the user selected.

Change-Id: I8b72146b7ee496d6c88539e7314a96f069f507fb