Fix crash with deleting expanded nodes
LT-15396
Deleting an expanded dictionary node results in crash

RefreshView creates a List expandedNodes in the tree.
Then the tree was recreated and some of the nodes in the List
expandedNodes might not be in the recreated tree which resulted
in a null value when trying to find the expandedNodes.

Change-Id: Ia457737d0949939157a8f9ffaa66328d8a14f550