LT-15894 & 15858 prevent LIFT import making duplicates
LIFT import had problems when importing a Ranges file where a
lexical-relation element specified a guid but not a sufficient
label for matching with an existing LexRefType that had that
guid. A duplicate object would be created. With this change, our first
choice is to match the guid. If that succeeds, we fill in any missing
attributes, and report any conflicts as errors.

I expected to have to fix similar problems for other range elements,
but there are no other cases I can find where we create one
using a specific guid from the LIFT file.

Change-Id: I2bb031a13a33f62c5b00c66eaea36a253bf75e33