LT-15695: Make creating pronunciation WS safer
Problem was caused by first-ever access in project to
DefaultPronunciationWritingSystem while LIFT import was
broadcasting changes. The method tries to create a WS if
one does not exist, but making changes is not allowed in this
state.
Fixed this to do something desperate one time if it happens,
while fixing it ASAP; also made a change that should ensure
this always happens safely.

Change-Id: Ief663d7609b49103b54c8f452e84e441f9dcee32