LT-15964, Improve Word 2007 interlinear export
Better handles morpheme-level rows with too few items.
Word seems to align things more reliably if we make sure
that a 3-column matrix has three elements on each row.
The change causes us to fill in any missing elements with
one containing just a space.

The empty element does not have the usual style for the row.
I don't think this will matter unless the user wants to fill
something in for the missing gloss. If that is needed we
could make it smarter, but I wanted to keep it as simple as
possible for a hot fix.

Change-Id: I55d5911e49dffce06a3c1d6fadc0b0100281ef5f