Allow Flex to reliably work with Keyman 9 (LT-15826)
This change implements the new IWindowsLanguageProfileSink
(that requires a change in Palaso, PR #190) to detect a change
of the input language instead of relying on the Windows
messages.

Thanks to Marc Durdin for the analysis of the problem.

Change-Id: Id5989fc7959da576e2e7d1764bb245a2816e4cad
(cherry picked from commit da89dd623a913ed737bc54b20e2d1868afb981d5)