Add remaining global File menu items
This change also introduces the new IArea and ITool interfaces
and it includes one do-nothing-yet implementation for the
'Lexicon' area (no tools yet). This allows the main window
to start using the IArea interface, even it not much is done.

While not much is done with the new interfaces, the code
now exists that figures out (Reflection) which IArea
implementations exist in the system. This should allow
users to select which areas and/or tools to install, beyond
any that are required, such as the Lexicon area.

Change-Id: I647c6b583e129b7b9476bf635253fecc99fb6447