Check if current form is enabled (LT-16452)
Previously when the user opened the find dialog while being in a
field with activated ibus all keyboard input in the find dialog also
ended up in the field in lexicon edit. This change checks if the
top level control is enabled before updating/committing text in the
lexicon edit field.

Change-Id: I3ac0e046abe659ef230f1416f5a30b9b9fb1d232