Move LexTextDll into LanguageExplorer
This includes moving the links to the xml configuration files
from the LexTextDll project into the LanguageExplorer. They
are removed from the project, as the files get removed. Having
them included in a project like this helps considerably
in the effort, since they can be searched like the code files.


Change-Id: I6984eade17b871c2a883d860115e2dbd317066a0