Properly reset pre-edit (LT-16671)
We can rollback if we have a different depth or if we have tasks
since our last mark. Without properly rolling back we our undo
task might be in a weird state so that when we try to save the
current record when switching to a different record we crash. This
change does a rollback of the pre-edit text thus removing the traces
we left in the action handler.

Change-Id: I01b2be56136c36631075e135f8cb284cbe8ebd1e