Fix crash when saving changes (LT-16673)
When a pre-edit window is still open after typing in the title of
a text and the user clicks on the button to insert a new text, we
save the changes of the current record by ending the undo task.
We used to call EndOuterUndoTask() which was not implemented and so
caused a crash. This change now calls EndUndoTask() instead.

The underlying problem that caused the crash was that when the user
clicks the insert text toolbar button we don't get a lost focus
event until after inserting the new text (at least on Linux), so the
keyboard handler doesn't have a chance to cleanup the pre-edit
before the changes get saved. This also has the side-effect that any
pre-edit text will get saved as well.

Change-Id: I3bbfde9e0da4d4da0eaaa88e9e5602c578a17eef