Work on default properties
This change loads the persisted properties
and then has the code make sure all default properties
are in the table. The change also removes any known obsolete
properties, if present in the persisted properties.

The defaults properties are set (if needed) by a newly exposed
method on IPropertyTable. This method had been in the older
system, but not part of the interface.
It is now part of the interface.

I also reworked the "RemoveProperty" methods on the interface
and implementation.

A cosmetic change was to add more region blocks,
where they were they were missing for IFlexComponent.

Change-Id: Icb861054cc105b3536801e9d6247a28e00e8cfbc