Fix a bug in validating a new project folder
The return should be outside the loop through accessRules, not inside
at the end of the loop.  This was causing valid project directories to
be rejected because the program couldn't tell that the user had read
and write access to them unless the first rule encountered belonged to
the user or to a group the user is in.

Change-Id: I4d2022474f90a91ec2d86c2e529c9de99b946023