Remove binaries from .NET framework from repo
We used to have a few assemblies from the .NET framework in our
repo. They got copied to the output directory during a build.
The likely reason we had these assemblies is that they are not
available on Linux (they are part of WPF which is not supported
on Mono). However these assemblies cause problems on Linux (e.g.
when restoring nuget packages) because the wrong assemblies get
picked up that don't work on Linux.

This change removes these assemblies and for Windows relies on
the ones that are ones from the framework installed on the
system. For Linux we exclude the code from compilation.

Change-Id: I7e2406b17e5bf46dfa94bac19eb32a8ef958efef