Allow to split building of localization
This changes adds a new property "Build" to the LocalizeFieldWorks
task that allows to build only the source files for the resource
dlls, only the binary resource dlls, or both.

The motivation behind this change is the desire to create the source
files only once when building packages which should cut away roughly
half the time needed to build localization.

Change-Id: I7ee3464a6ea513a41da8ac96fcce71a46fc70404