Use ICU_DATA environment variable to get data directory
This change makes COMInterfaces more independent of WinForms. There
are more changes necessary, but it is a step in that direction.

Previously we read the ICU data directory from the registry. Now
the FieldWorks class (or when running tests SetIcuDataEnvironmentVar
class) reads the registry and sets the ICU_DATA environment variable.

Change-Id: I12a5c6d3d4337369e73cc434197dedae8fec31b0