Handle Guid.Empty in PoSFactory.Create(Guid, IPartOfSpeech)
PartOfSpeechFactory has two Create() methods that take two parameters,
a GUID and an owner. The version that takes an ICmPossibilityList as
the owner will, if passed Guid.Empty as its GUID value, call Create()
with no parameters in order to get a real GUID in the new PartOfSpeech
object. But the version of Create() that takes an IPartOfSpeech as the
owner was behaving differently: if passed Guid.Empty, it would create
a PartOfSpeech with Guid.Empty as its GUID. This commit fixes that
behavior. Now both versions of PartOfSpeechFactory.Create() will
behave symmetrically, assigning a real GUID if passed Guid.Empty.

Change-Id: Idfabd7061563eb03e2f1955bf1949781fd4314da