Expose Microsoft.Practices.ServiceLocation in pkg-config
Some parts of FDO require Microsoft.Practices.ServiceLocation.dll.
However, the version we're currently using (1.0.0) is not available
as nuget package, so we expose the assembly to pkg-config so that
projects that use FDO can use it. This is a hack - the proper way
will eventually be to update Microsoft.Practices.ServiceLocation
to a newer version with a nuget package and to reference the nuget
package in the projects that use FDO.

Change-Id: I99c6a16ec61ebeefd0b15ec058242edf0ad4000f