Re-enable Linux package builds on Debian
Use of the home directory in package builds is now forbidden, by
setting HOME to /nonexistent. FW builds do use the home directory
in a couple of places so it's necessary to fix these.

One place is due to a bug (passing an empty value to `cd`) and the
other is due to setting up keys in HKCU, which is inappropriate when
building a package. We now set a property when package building,
so that targets can be made conditional on it.

Change-Id: I88ea085eb20c801427c3023e97567d701ab1eac4