Cleanup DetailsView Layout issue
When Right-clicking to open Dictionary Configuration in Lexical
entry, the view goes directly to the item (ie example, headword)
the initialization in this case changed the behaviour, so we need
to add a comparison for this.
The default width of the DictionaryConfigurationDlg is too narrow
causing hidden buttons in the detailsview. Setting a MinimumSize
for the dialog that will fix this problem. (LT-16992)

Change-Id: Ie85f9f0fd3ae1def0825373324deff5389cf514b