Remove circular reference issues between LangExp & FieldWorks
This work is need to remove future circular references between
Language Explorer & FieldWorks and xWorks and FdoUi, when those
two assemblies get assimilated into Language Explorer.

There are some tests thar make sure the new Reflection code works.
If the classes/method change or are moved, the idea os those new
tests will point that out at build time, not later at runtime.

A secondary change was to renamed the two new parameter object
classes introduced in the last round of work. This rename
affected most of the files that are changed in this commit.

Change-Id: Ie5b936eef2308c25cdcc3bed9319bd01dcc937f7