Allow choosing copies of reversal index configurations
Fixes LT-17170.

* For the list of reversal indexes that can be chosen in the reversal
index area or the bulk edit reversal entries area, make the list be
the set of reversal index configurations rather than the list of
analysis writing systems and the GUIDs of the associated sets of
reversal index entries.
* When a reversal index configuration is chosen, also set the GUID for
the set of reversal index entries, based on the writing system of the
reversal index configuration.
* Don't show an choice for All Reveral Configurations in the drop-down
list in the reversal index area or the bulk edit reversal entries
area.
* Moved some methods to a utility class so they can be used outside of
XhtmlDocView.

There is now a regression: Blank page shown
when first change reversal index configuration.

Change-Id: I36f38519a0d3a357bf738ebe5943dad1541e5c91