Revert "Fix package"
This reverts commit 650195de05cf14e112d7560d9ee1bfeb978de83b.
It turns out the reason why the files didn't get removed previously
was a problem with our Jenkins build job so that the jobs built
with an outdated commit from March. Now that this is solved I can
revert this commit. However, we still want to remove LibChorus.dll*
from the package.

Change-Id: I39d01660b40e3f3db78dfddfa6b16c64d7dde59e