Display exception type when we can't save fwdata file
When saving the fwdata file fails we catch all exceptions. However,
so far we didn't output what exception we got. This is good for the
end user but less helpful for a developer trying to see what went
wrong. This change adds the exception type after the error message.

It also retrieves the application name dynamically so that when FDO
is used as library from other products the name will show up
correctly.

Change-Id: I89e5745ecca1843949491bde111af510f821cc3e