Fix a few bugs in TsStrBldr.SetProperties
SetProperties() would do the wrong thing in several scenarios, mostly
when replacing text properties in the middle of a run. (E.g., if you
have "abc" and try to replace the text properties of just "b", it
would end up with an invalid state, like a run containing "a" and a
second run containing "b" but no run containing "c".) Fixed those bugs
and wrote unit tests to cover them.

Change-Id: I96de9e9b98433f8f297771e6696c792afc30737b