Fix crash on exiting FieldWorks
When ThreadHelper.Dispose() is called, it tries to dispose the
WinForms control that it held onto, using Invoke() to try to ensure
that the call will be called from the main thread. But Invoke() uses
the Invoke method of the WinForms control being disposed, so the
control is invoking itself inside its Dispose() method, like Ouroboros
eating its own tail. Unsurprisingly, that fails.

Change-Id: I4614a0d4d3611c505749729e1080fa248e08e320