Don't skip saving following Exception checking time
If there is an exception checking whether a key is down (to determine
last activity time), swallow it and return the last recorded activity
time. This will prevent indefinite delays of saves and Dictionary
preview refreshes if we get in a bad state. The worst thing that can
happen is the preview flickers while the user is holding a key down.

Change-Id: Ic1196f4bf44777ca9d1d45a6fe4582fe1d6590fc