LT-7476 partial: increase size in Morph Type chooser
This really needs a stylesheet applied to it (and using the
"Normal" style, like other combos with analysis WS in them), as
was done in LT-7389 for the Sense Choosers (LinkEntryOrSenseDlg.cs).
But I don't know how to do that, so brute-forcing it to a slightly
larger size makes it less glaring than what it has been.

Change-Id: I464ad5a49cd78fc587891f1c1fb033fe199404ec