[uce] Display errors when saving PUA characters
Previously we blatantly claimed that ICU files are locked even
when running gennorm2 produced a different error. That hid the
real problems.

Also fix `environ` to include the path to `gennorm2` - seems
the location where it gets build got changed a while ago.

Also don't include a manifest file when compiling in VS - this
matches the setting in FieldWorks.csproj and allows to debug
the app on Windows.

Change-Id: I4a45261d557d23a6b22da648711bed8fad6ad974