LT-18295 (Partial): Enable default WS for Gloss (or SumDffn)
That the default WS is disabled in the config file should not be
a problem. The greater problem that this exposes is that we do not
sync and validate all parts of the config on loading, but wait until
the individual nodes are loaded into the configuration file. That
will be fixed in a follow-on commit.

Change-Id: I4ac7e6ce0109639ff5652080b27ef2e0888f5f05