[uce] Fix Icu54DataDir registry setting (LT-14263)
With ICU 54 the data directory is the same as the normal ICU
directory. It seems this registry entry doesn't matter for the
rest of FW but for UCE it needs to point to the correct directory.
The `setup-user` script gets run at each start of FW and UCE and
among other things restores registry settings.
(cherry picked from commit 88a52f4f7a8bd0db942b7740bf57b64b11ce4539)

Change-Id: I2c4be59cabe540385233c6b178d598ce6f8823d8