LT-18460: Handle dynamic loading of ICU in icu.net
icu.net now loads the highest available icu library. This causes
failures on Xenial which comes with ICU 55. This change limits
the ICU library version to 54 so that our custom ICU gets
loaded.

Because of namespace conflicts we have to specify the full
namespace for the ICU classes in COMInterfaces. This is ugly,
but should only be necessary for a short time until we got rid
of the ICU classes in FW.

See https://gerrit.lsdev.sil.org/#/c/4616/
and https://gerrit.lsdev.sil.org/#/c/4816/ for previous reviews
of this change.

Change-Id: I22e6a622c482465b380b2753ac5c0b446678e155