LT-18449 - Have factory methods on areas/tools create clerks
The record clerk repository will have the given factory method
create a clerk, if it is not already in the repository.

The areas/tools know how to create clerks of interest using:

"Func<LcmCache, FlexComponentParameters, string, RecordClerk>"

The factory method will throw an exception if asked to create a clerk with
an Id it does not know how to create.

Change-Id: Ia5bc2c7144c221e64d48aaf5631fd22fc24816d9