Revert https://trello.com/c/X3q9oiRv (Dependency Download)
TeamCity was recently changed to conceal potentially-
sensitive data from guests. Chrisvire's BuildUpdate now
requires credentials to read dependencies from TeamCity,
but it would be a hassle to supply credentials on the
build server, so we're removing that functionality for now.

Perhaps in the future we can create a script that runs
locally to read dependencies from FLExBridge on TeamCity,
then check in the results.

Additional change in this commit:
 - fix a typo
 - request user input as soon as possible in developer build scripts

Change-Id: I8b7ff29fbf1cbc498cb53521212d0ec2d4f4c2a9