LT-18449 - Get XmlDocConfigureDlg to using ConfigureHomographDlg again
Gory details: The old Mediator command "CmdConfigHeadwordNumbers" had been
used by a sub-menu in the Tools->Configure menu. But, for the past couple of
years, or so, that menu was never visible. or enabled. (Cf. Now removed
method "OnDisplayConfigureHeadwordNumbers" in DictionaryConfigurationListener.)
But, XmlDocConfigureDlg still fired off a message to the Mediator to get the
ConfigureHomographDlg to do its thing. So, this change made it easier,
and moved ConfigureHomographDlg into the same namespace as XmlDocConfigureDlg
and created and ran ConfigureHomographDlg directly.

The bogus Tools->Configure->Headword Numbers... menu is now gone.

Change-Id: I988fa22b3aa201b09c511cac26e1f5d9654828a2