Rearrange setup-user registry key logic
This unfortunately increases the nesting of the registry key
upgrading logic.

But it provides more structure to a process that has caused some
trouble recently.

It does also avoid a potential problem, which is that a user who
installs FW 9 afresh would have had empty settings written into FW 8
registry keys. That shouln't normally be a problem, but it would be
unfortunate for that to cause trouble in a special case with a user,
or with a tester who installs FW 9 and then uninstalls and installs
FW 8.

Change-Id: I7a3f0bab654d45a1533429c06ca01edae36f2fd7