LT-18449 - Add IRecordClerk and IRecordList interfaces
This is the first of several rounds of work that aim to get rid of
RecordClerk in favor of RecordList. The class docs for RecordClerk say:
"This class, essentially, adapts a RecordList to the xCore/xWorks environment."
If that is true, then RecordClerk isn't needed in an environment that is
free of XCore.

Change-Id: If6db325cec73246c77e34f9044331fff6b44edbd