Keep IServiceLocation public when packing FDO
When running ILRepack on FDO we want to keep the classes and
interfaces in `Microsoft.Practices.ServiceLocation.dll` public.
This is necessary because IFdoServiceLocation derives from
IServiceLocation. While it compiles without this change on older
Mono versions, it stopped working with Mono 5 because the compiler
sees IServiceLocation as internal, thus making it impossible to
access any method on that interface.

Change-Id: Id15d3829dd9217a69374952ce4164c450f4146b1