LT-18638 & LT-18639: Get Delete to work (sort of).
This change enables/disables the Edit-Delete menu and related toolbar button.
The text of the menu item changes to match the class of the deleted item,
as well as the tooltip of the delete toolbar button.

The part that does not work is the actual deletion,
which for some reason fails in UOW-land in LCM.

Change-Id: Ic9de038e9b49d0637e3ce004366a984eecbc590f