Adjust for Mono 5
This removes `#if __MonoCS__` which isn't defined on Mono 5 because
it uses the Roslyn compiler. Instead we check the platform at runtime.

This also includes a hack to force msbuild to compile with gtk-sharp
2.12. Otherwise it uses 3.0 which fails to build (at least when
compiling in Rider), and fails to run. An additional benefit of this
change is that we don't need conditionals and can compile the same
code on Linux and Windows.

Change-Id: I589e0384e2d81f15f2b54e313f584f62e42596e9
(cherry picked from commit 8eb91854aa3aecf458d59f434b12e3b0b8a871b3)