Fix mistake in environ
environ needs to be sourced. This line implies it can be run directly,
which will lead to confusion.

It is possible some system out there relies on this #! line. If so,
changing this will help us find it and correct that system to source
this file instead. (Or I will be proved wrong and we need to restore
this line.) Although the file is not executable, so the #! is probably
not being used for execution.

Change-Id: I2e975259d197fdf8d194a64b94773c8848633876