LT-19499: Get Problems Tool to work
Strictly speaking, this was more like: "Sew the patient back up,
since the condition is 'terminal'."

One of the two slices was not implemented. The second was interested in
adding new dictionary entries in a vernacular language dictionary,
for words in a "Text" that were in English that described the problem.

Change-Id: I42c9e1023086997608b6f2347394ae6a95bb32f6