Fix LT-20221: Refactor RenderEngineFactory
* Store Uniscribe engines and graphite engines
  in the same cache

The bug was caused by a corrupt font table being returned
from a VwGraphics object when trying to create a GraphiteEngine.

If the corrupt data happened to look enough like a valid graphite
font table it would create a GraphiteEngine that would render
garbage and store it for use with the given writing system.

This refactor is an effective band-aid to the bad data as it will
not repeatedly try to create a graphite font. The source of the
corrupt font table proved too difficult to determine.

Change-Id: I0ffe2455d327cd94ac6597a6b18f0200bc76915f