Revert "Update Anthropology Category 614 to match OCM"
This reverts commit 91ac963052d5044e26cef3d741ee0b9489739e3f.

Reason for revert: The old name used to be correct.
This category was changed from Sibs to Clans, and category 618
was change from Clans to Localized Kin Groups, and their
scope notes were rewritten in 2005. There have been 44 changes
to OCM in 1997, 2005, 2012, and possibly other years, and verifying
these updates is not critical to our impending stable release.
On the other hand, a major release such as this one seems the most
opportune time to make these changes, but they should all be
made at once.
https://hraf.yale.edu/resources/reference/outline-of-cultural-materials/#id618

Change-Id: Icb23c6cff9bf2e73a15f97983221459a7a99386b