LT-19289: Fix keyboard switching on Bionic GnomeShell
For some reason when switching keyboards we get a kill focus
event which caused us to activate the default keyboard. This
resulted in toggling between the default keyboard and the keyboard
of the current field. This change fixes the bug by skipping the
first kill focus event after we switched keyboards.

A downside of this fix is that we won't switch to the default
keyboard if the user clicks in a new field in FW and then activates
another app within 500ms.

Change-Id: I299a96ed78e2cb3c252051f219988267013d631f