LT-19915: Fix Word 2019+ Export
Remove the <wtblGrid> elements from the exported Word XML file.
These elements cause MS Word 2019 and later to fail upon opening
the XML file with an error message:

"We're sorry.  We cannot open the document because we found a problem
with its contents.  Not enough memory resources are available to
complete this operation."

Removing these elements fixes the error and allows users of
Word 2019 and later to open the Discourse Chart Word XML file.
Removing the elements does not appear to have an adverse effect
on the chart output.

Change-Id: Ib0f3b204ec67083c593eab96751944e6b2718cea