LT-20660: Fix Crash when changing morph type
When the user changes the type from stem to affix (or affix to stem)
we dispose of some of the slices and new slices are created. In this
case we do not want to call UpdateDisplayFromDatabase() since it will
also try to dispose of slices, resulting in the exception that was
reported.

The call to UpdateDisplayFromDatabase() was added with Commit e53014aa
to address the refresh problem described in LT-20447.

Moving the UpdateDisplayFromDatabase() call into MakeMorphTypeChange()
has two benefits:
1. It is included in existing logic that already treats a change from
   one type to another as unique.
2. It fixes a refresh problem that is similar to what was reported in
   LT-20447. This problem existed if the user changed the morph type
   (to another type in the same stem/affix grouping) by using
   Auto-Complete instead of displaying the ‘Choose Morpheme Type’
   dialog. After the change was completed the ‘Morph Type’ was only
   displaying the characters the user entered, not the complete name.
   With this change we now update the display regardless of how the
   morph type is changed (if it is changed to another type in the same
   stem/affix grouping).

Change-Id: Ibcbb82d44ccb4091286a9461d02cb369575f6076