LT-20613: Preserve the order of fields
The order of custom fields and the order of the writing systems were
lost after using the Configure Interlinear dialog. The order is now
preserved.

Know Problem: When the user renames a Custom Field then its order and
the option to display it will not be preserved.  It will default to
not being displayed and will appear at the end in the ‘Configure
Interlinear Lines’ dialog.

Change-Id: If29afe6bbfc32f4669ad00f149ff13b4e6d402a5