LT-20846: Display default Gloss instead of stars
When selecting a different parse from the combo box we now always
pass true to LoadRealDataIntoSec() for fLookForDefaults. This makes
the behavior described in this defect consistent with the behavior
when the user switches from the Baseline tab to the Analyze tab.
Handle_AnalysisChosen() only appears to be called from the combo box
so I did not try to pass in a fLookForDefaults flag. I don’t think
it is necessary.

Change-Id: I904266c9552cc2aee3af9267db880a413b552f4b