LT-20794: Fix Linux crash after importing data
The import wizard had a ‘using’ statement on m_app.SettingsKey,
which resulted in the settings key being marked for delete when
we exit the using block. On Linux this resulted in a crash when we
later tried to access the settings key for write in
FieldWorks.FieldWorks_OnSave() at this line:
    settings.LatestProject = e.Cache.ProjectId.Handle;

The exception we were getting was:
System.IO.IOException: Illegal operation attempted on a registry
key that has been marked for deletion.
"at Microsoft.Win32.RegistryKey.CreateMarkedForDeletionException
() [0x00000] in /build/mono5-sil-5…"

There is similar code in the same method (and in many other
files) where there is a ‘using’ statement, but instead of
using the settings key it is using a created subkey:
using (var key = m_app.SettingsKey.CreateSubKey("ImportDictFiles"))

Possibly the using statement on m_app.SettingsKey was
a cut and paste error.

Note: The import wizard code around the area of the using
statement on m_app.SettingsKey hasn’t changed since the
file was added to Git so it’s unclear how this worked in
earlier releases.

Change-Id: I7aa2b0bc798b3f833133df34c3957a36fd7ce038