LT-20837: Fix empty text so that it uses DefaultVernWs
When restarting Flex, and there is a non-English Flex UI, then
we were using the UI ws for an empty text instead of using the
DefaultVernWs.
There was existing code that would check for an empty text
and change the ws to DefaultVernWs, but instead of checking
against userWs it was checking against English (the global
default).  So the code was only being hit when English was the
userWs.
Note: The default is set to Cache.WritingSystemFactory.UserWs in
liblcm\src\SIL.LCModel\DomainImpl\GeneratedClasses.cs
StTxtPara:Contents:get

Change-Id: Ic06e0b15b0e4d7125c9e52e203d49b297e671ee7