LT-20880: Fix FLEXtext exports
Changes in InterlinearExporter.cs:
We were writing the wrong ‘lang’ value.  Change the code to
use the WriteItem() method to write the complete element.
This is the same method we use to write other similar data.

Changes in InterlinearExportDialog.cs:
Fix the following exception that only happens when a debugger
is attached:
    Class: System.InvalidOperationException
    Msg: Cross-thread operation not valid: Control 'm_exportList'
         accessed from a thread other than the thread it was
         created on.
Here is an explanation why it only throws when a debugger was
attached:
https://stackoverflow.com/questions/3972727/why-is-cross-thread-
operation-exception-not-thrown-while-running-exe-in-bin-debu

Change-Id: Ie01902631b1b4e64f02c0466fea9c5c5d6e4d199