LT-20651: Prevent creation of null files
This change, along with the related changes in Palaso, are an
attempt to prevent the creation of null files. The theory is that
FieldWorks is closing and power is lost before the operating
system cache is written to disk. These changes cause the cache to
be flushed immediately after writing instead of relying on the
cache manager lazy writing mechanism.
https://docs.microsoft.com/en-us/windows/win32/fileio/file-caching

Change-Id: I2dfb657c6dc579eb8e7a9e1b4956eb3c3e1c1864