LT-20962 Bump DialogAdapters to fix crash
DialogAdapters is crashing in Ubuntu 20.04 (in or out of flatpak) with
System.InvalidOperationException
...
at GLib.Idle.Remove
at DialogAdapters.CommonDialogAdapter.CommonTearDown

DialogAdapters received a fix for this. And also since the version we
were using, it has split into two packages, for Gtk3 and Gtk2. This
change also specifies the package name to use Gtk2.

Change-Id: I43e287f2e37245804e0ddc4472a4fe7c9e1c9dfd