LT-20768: Prepopulate new text charting templates
* Prepopulate when creating from the charting tool, too.
* Name new template items.

Storing the names and structure in Localized Lists XML was considered,
but that would have required shipping LocalizedLists.en.xml, and
loading the portion could have presented a pain.
Using the UI language WS was considered, but if it is not in the
project, the data would be hidden. Using a UI WS run in the default
analysis WS Name string would keep the entire string in the UI WS
until FLEx is restarted.
UI language in a run of the default analysis WS should be fine as long
as the analysis font supports the UI alphabet.

Change-Id: Iaf02f81f265e0b3fc0c7a024881ac77f32aa5e35