LT-20930: Fix custom field links to other projects
XmlBrowseViewBaseVc.DoHotLinkAction() was designed to call the base
class if the FollowLink call threw an exception. But we were
catching the exception, displaying an error message, and not
checking the return value for failure. This change checks the return
value so we will call the base class on a failure and also adds a
property so we can suppress the error message.

Change-Id: Id16157521f98ef5d1559e43cbe1f46dfad1dac82